Attack Update
Started on two fighting functions. The goFight function is used to fight a boss and will make the boss stats appear on the screen along with the boss name.
The attack function is used by both the user and the boss. when used by the boss it will have specific dialogue as compared to the user which will display what weapon youre using.